remove elasticsearch mock Signed-off-by: TOURI ANIS <anis-1.touri@rte-france.com>
