+++ src/main/java/com/powsybl/network/conversion/server/elasticsearch/ESConfig.java
@@ -8,10 +8,8 @@ package com.powsybl.network.conversion.server.elasticsearch;
 
 import org.elasticsearch.client.RestHighLevelClient;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.data.elasticsearch.client.ClientConfiguration;
 import org.springframework.data.elasticsearch.client.ClientConfiguration.TerminalClientConfigurationBuilder;
 import org.springframework.data.elasticsearch.client.RestClients;
@@ -29,7 +27,6 @@ import java.util.Optional;
 
 @Configuration
 @EnableElasticsearchRepositories
-@Lazy
 public class ESConfig extends AbstractElasticsearchConfiguration {
 
     @Value("#{'${spring.data.elasticsearch.embedded:false}' ? 'localhost' : '${spring.data.elasticsearch.host}'}")
@@ -46,18 +43,6 @@ public class ESConfig extends AbstractElasticsearchConfiguration {
 
     @Value("${spring.data.elasticsearch.password:#{null}}")
     private Optional<String> password;
-
-    @Bean
-    @ConditionalOnExpression("'${spring.data.elasticsearch.enabled:false}' == 'true'")
-    public EquipmentInfosService studyInfosServiceImpl(EquipmentInfosRepository equipmentInfosRepository) {
-        return new EquipmentInfosServiceImpl(equipmentInfosRepository);
-    }
-
-    @Bean
-    @ConditionalOnExpression("'${spring.data.elasticsearch.enabled:false}' == 'false'")
-    public EquipmentInfosService studyInfosServiceMock() {
-        return new EquipmentInfosServiceMock();
-    }
 
     @Bean
     @Override
