+++ src/main/java/com/powsybl/network/conversion/server/elasticsearch/EquipmentInfosRepository.java
@@ -7,8 +7,6 @@
 package com.powsybl.network.conversion.server.elasticsearch;
 
 import com.powsybl.network.conversion.server.dto.EquipmentInfos;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 import org.springframework.lang.NonNull;
 
@@ -17,8 +15,6 @@ import java.util.UUID;
 /**
  * @author Slimane Amar <slimane.amar at rte-france.com>
  */
-@ConditionalOnExpression("'${spring.data.elasticsearch.enabled:false}' == 'true'")
-@Lazy
 public interface EquipmentInfosRepository extends ElasticsearchRepository<EquipmentInfos, String> {
     Iterable<EquipmentInfos> findAllByNetworkUuid(@NonNull UUID networkUuid);
