--- src/test/java/com/powsybl/network/conversion/server/EquipmentInfosServiceMockTests.java
@@ -1,51 +1,0 @@
-/*
-  Copyright (c) 2021, RTE (http://www.rte-france.com)
-  This Source Code Form is subject to the terms of the Mozilla Public
-  License, v. 2.0. If a copy of the MPL was not distributed with this
-  file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.powsybl.network.conversion.server;
-
-import com.google.common.collect.Iterables;
-import com.powsybl.iidm.network.IdentifiableType;
-import com.powsybl.network.conversion.server.dto.EquipmentInfos;
-import com.powsybl.network.conversion.server.dto.VoltageLevelInfos;
-import com.powsybl.network.conversion.server.elasticsearch.EquipmentInfosService;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-import java.util.List;
-import java.util.Set;
-import java.util.UUID;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * @author Slimane Amar <slimane.amar at rte-france.com>
- */
-@RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE, properties = {"spring.data.elasticsearch.enabled=false"})
-public class EquipmentInfosServiceMockTests {
-
-    private static final UUID NETWORK_UUID = UUID.fromString("38400000-8cf0-11bd-b23e-10b96e4ef00d");
-
-    @Autowired
-    private EquipmentInfosService equipmentInfosService;
-
-    @Test
-    public void testAddDeleteEquipmentInfos() {
-        List<EquipmentInfos> infos = List.of(
-                EquipmentInfos.builder().networkUuid(NETWORK_UUID).id("id1").name("name1").type(IdentifiableType.LOAD.name()).voltageLevels(Set.of(VoltageLevelInfos.builder().id("vl1").name("vl1").build())).build(),
-                EquipmentInfos.builder().networkUuid(NETWORK_UUID).id("id1").name("name1").type(IdentifiableType.LOAD.name()).voltageLevels(Set.of(VoltageLevelInfos.builder().id("vl2").name("vl2").build())).build()
-        );
-
-        equipmentInfosService.addAll(infos);
-        assertEquals(0, Iterables.size(equipmentInfosService.findAll(NETWORK_UUID)));
-
-        equipmentInfosService.deleteAll(NETWORK_UUID);
-        assertEquals(0, Iterables.size(equipmentInfosService.findAll(NETWORK_UUID)));
-    }
-}
